bug: simple-mobile-backend example app
The app that's published to the Serverless Application Repository has
replaced the custom policy with a policy template. The current published
version also has a bug in it where it is not Ref'ing the parameter. This
commit updates the template to match what's in the Serverless
Application Repository and also fixes the Ref bug.