feat: hard-code list of managed policies and use that by default
This was the only AWS Service call in SAM. By removing this it allows for the SAM transformer to be run offline.

Maintained the initial code with an option so that in the future we an expose a way to grab the authoritative list from IAM. This also allows us to easily grab an updated list periodically.