Refactor hash generation in API Deployment's logical ID
In current implementation, due to use of str(swagger) and json.dumps(domain),
the hash output depends on Python's default iteration behaviour and can result
in inconsistency even if swagger/domain did not change.
This PR ensures the hash input is a dict object, as LogicalIdGenerator will
produce consistent hash if the input is a dict object.