feature(kms): Allow Functions to Decrypt Environment Variables
...If They Have Been Encrypted

When a Function specifies that its environment variables should be
encrypted by supplying a value for `KmsKeyArn`, a corresponding
Policy will be generated for that Function's execution Role which
allows it to decrypt the environment variables with that key.
Permissions are restricted so that it should be able to do nothing
else with the key.