Release/v1.37.0 (#2069)
* Revert "fix: Crash when using an invalid method in open api (#2001)" (#2021)

This reverts commit d57b1329e1b99169f2d5ca59c1c41eee66b33750.

* fix: Increase PageSize of ListPolicies Paginator (#2033)

SAM runs within a Lambda function and loads IAM Managed Policies once per
Lambda. Previous to this, SAM would call IAM 9 times which could cause
throttling by IAM. With this change, we update the MaxItems from the
default of 100 to the max (1000). In local testing, this has shown
a 0.6 second reduction in the latency in calling IAM.

Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>

* Revert "Issue 1508 remove check requiring identity ... (#1577)" (#2038)

This reverts commit 0eb3630410ae569503513e52670ece61dc4d4301.

This change caused regression, reverting until the bug is fixed.

bug: `ReauthorizeEvery` can be a `dict` when intrinsic functions are used.

* chore: bump version to 1.37.0 (#2068)

Co-authored-by: Jacob Fuss <32497805+jfuss@users.noreply.github.com>
Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>
Co-authored-by: _sam <3804518+aahung@users.noreply.github.com>