chore: clean up cfn-lint ignored files
When we have new features, cfn-lint might not have support to all those
features yet, so we usually just ignore some files until that's supported
but then we rarely actually remove that so the files just remain ignored
forever. So this is an overall cleanup to remove all the ignores that are
not needed.

The next step would be to confirm why all these files are ignored, and
eventually fix some that make sense to fix (some have errors on purpose).